#98 메인, 마이페이지 - module 저장, 불러오기
