+++ src/main/java/com/example/lazier/dto/user/MemberModuleResponseDto.java
@@ -1,5 +1,6 @@
 package com.example.lazier.dto.user;
 
+import com.example.lazier.persist.entity.user.ModuleYn;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -10,5 +11,27 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @Builder
 public class MemberModuleResponseDto {
-	private String userModuleList;
+
+	private boolean weatherYn;
+	private boolean exchangeYn;
+	private boolean stockYn;
+	private boolean newsYn;
+	private boolean youtubeYn;
+	private boolean quoteYn;
+	private boolean todolistYn;
+	private boolean workYn;
+
+	public static MemberModuleResponseDto of(ModuleYn moduleYn) {
+		return MemberModuleResponseDto
+			.builder()
+			.weatherYn(moduleYn.isWeatherYn())
+			.exchangeYn(moduleYn.isExchangeYn())
+			.stockYn(moduleYn.isStockYn())
+			.newsYn(moduleYn.isNewsYn())
+			.youtubeYn(moduleYn.isYoutubeYn())
+			.quoteYn(moduleYn.isQuoteYn())
+			.todolistYn(moduleYn.isTodolistYn())
+			.workYn(moduleYn.isWorkYn())
+			.build();
+	}
 }
