+++ src/main/java/com/example/lazier/dto/user/MemberModuleSaveRequestDto.java
@@ -1,10 +1,14 @@
 package com.example.lazier.dto.user;
 
 import javax.validation.constraints.NotBlank;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+@Builder
 @Data
+@AllArgsConstructor
 @NoArgsConstructor
 public class MemberModuleSaveRequestDto {
 
@@ -12,5 +16,19 @@ public class MemberModuleSaveRequestDto {
 	private String userId;
 
 	@NotBlank
-	private String userModuleList;
+	private boolean weatherYn;
+	@NotBlank
+	private boolean exchangeYn;
+	@NotBlank
+	private boolean stockYn;
+	@NotBlank
+	private boolean newsYn;
+	@NotBlank
+	private boolean youtubeYn;
+	@NotBlank
+	private boolean quoteYn;
+	@NotBlank
+	private boolean todolistYn;
+	@NotBlank
+	private boolean workYn;
 }
