+++ src/main/java/com/example/lazier/persist/repository/ModuleYnRepository.java
@@ -1,0 +1,12 @@
+package com.example.lazier.persist.repository;
+
+import com.example.lazier.persist.entity.user.ModuleYn;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
+import org.springframework.data.repository.query.Param;
+
+public interface ModuleYnRepository extends JpaRepository<ModuleYn, Long> {
+
+	@Query("select my from ModuleYn my where my.lazierUser.userId = :userId")
+	ModuleYn findAllByUserId(@Param("userId") Long userId);
+}
