+++ src/main/java/com/example/lazier/service/module/NewsUserService.java
@@ -10,7 +10,7 @@ import com.example.lazier.persist.entity.module.NewsUser;
 import com.example.lazier.persist.entity.user.LazierUser;
 import com.example.lazier.persist.repository.NewsPressRepository;
 import com.example.lazier.persist.repository.NewsUserRepository;
-import com.example.lazier.service.user.MemberService;
+import com.example.lazier.service.user.MyPageService;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
@@ -27,7 +27,7 @@ public class NewsUserService {
 
 
   private final NewsUserRepository newsUserRepository;
-  private final MemberService memberService;
+  private final MyPageService myPageService;
   private final NewsPressRepository newsPressRepository;
 
   /**
@@ -39,7 +39,7 @@ public class NewsUserService {
   public void add(HttpServletRequest request) {
     long userId = Long.parseLong(request.getAttribute("userId").toString());
     //멤버인지 확인 -> 아닐경우 멤버서비스에서 Error Throw
-    LazierUser lazierUser = memberService.searchMember(userId);
+    LazierUser lazierUser = myPageService.searchMember(userId);
 
     // 뉴스 레포에 정보가 없을 경우 초기화 정보 생성
     if (!newsUserRepository.existsByLazierUser(lazierUser)) {
@@ -55,7 +55,7 @@ public class NewsUserService {
   public List<NewsPressDto> showUserPressList(HttpServletRequest request) {
     long userId = Long.parseLong(request.getAttribute("userId").toString());
     //멤버인지 확인 -> 아닐경우 멤버서비스에서 Error Throw
-    LazierUser lazierUser = memberService.searchMember(userId);
+    LazierUser lazierUser = myPageService.searchMember(userId);
     NewsUser user = newsUserRepository.findByLazierUser(lazierUser)
         .orElseThrow(() -> new UserNotFoundException("해당모듈 사용자가 아닙니다."));
 
@@ -67,7 +67,7 @@ public class NewsUserService {
   public void update(HttpServletRequest request, NewsUserInput userInput) {
     long userId = Long.parseLong(request.getAttribute("userId").toString());
     //멤버인지 확인 -> 아닐경우 멤버서비스에서 Error Throw
-    LazierUser lazierUser = memberService.searchMember(userId);
+    LazierUser lazierUser = myPageService.searchMember(userId);
 
     NewsUser user = newsUserRepository.findByLazierUser(lazierUser)
         .orElseThrow(() -> new UserNotFoundException("해당모듈 사용자가 아닙니다."));
