+++ src/main/java/com/example/lazier/service/module/UserQuotesService.java
@@ -7,7 +7,7 @@ import com.example.lazier.exception.QuotesException.UserNotFoundException;
 import com.example.lazier.persist.entity.module.UserQuotes;
 import com.example.lazier.persist.entity.user.LazierUser;
 import com.example.lazier.persist.repository.UserQuotesRepository;
-import com.example.lazier.service.user.MemberService;
+import com.example.lazier.service.user.MyPageService;
 import java.util.Optional;
 import javax.servlet.http.HttpServletRequest;
 import lombok.AllArgsConstructor;
@@ -19,11 +19,11 @@ import org.springframework.transaction.annotation.Transactional;
 public class UserQuotesService {
 
     private final UserQuotesRepository userQuotesRepository;
-    private final MemberService memberService;
+    private final MyPageService myPageService;
 
     public void add(HttpServletRequest request, UserQuotesInput parameter) {
         long userId = Long.parseLong(request.getAttribute("userId").toString());
-        LazierUser lazierUser = memberService.searchMember(userId);
+        LazierUser lazierUser = myPageService.searchMember(userId);
         // 중복 아이디 예외
         if (userQuotesRepository.existsByLazierUser(lazierUser)) {
             throw new UserAlreadyExistException("사용자 정보가 존재합니다.");
@@ -37,7 +37,7 @@ public class UserQuotesService {
 
     public UserQuotesDto get(HttpServletRequest request) {
         long userId = Long.parseLong(request.getAttribute("userId").toString());
-        LazierUser lazierUser = memberService.searchMember(userId);
+        LazierUser lazierUser = myPageService.searchMember(userId);
 
         Optional<UserQuotes> optionalUserQuotes = userQuotesRepository.findByLazierUser(lazierUser);
 
@@ -47,7 +47,7 @@ public class UserQuotesService {
     @Transactional
     public void update(HttpServletRequest request, UserQuotesInput parameter) {
         long userId = Long.parseLong(request.getAttribute("userId").toString());
-        LazierUser lazierUser = memberService.searchMember(userId);
+        LazierUser lazierUser = myPageService.searchMember(userId);
 
         UserQuotes userQuotes = userQuotesRepository.findByLazierUser(lazierUser)
             .orElseThrow(() -> new UserNotFoundException("사용자 정보가 존재하지 않습니다."));
@@ -57,7 +57,7 @@ public class UserQuotesService {
 
     public void delete(HttpServletRequest request) {
         long userId = Long.parseLong(request.getAttribute("userId").toString());
-        LazierUser lazierUser = memberService.searchMember(userId);
+        LazierUser lazierUser = myPageService.searchMember(userId);
 
         UserQuotes userQuotes = userQuotesRepository.findByLazierUser(lazierUser)
             .orElseThrow(() -> new UserNotFoundException("사용자 정보가 존재하지 않습니다."));
