+++ src/main/java/com/example/lazier/service/module/WeatherService.java
@@ -6,7 +6,7 @@ import com.example.lazier.persist.entity.module.Weather;
 import com.example.lazier.persist.entity.user.LazierUser;
 import com.example.lazier.persist.repository.WeatherRepository;
 import com.example.lazier.scraper.NaverWeatherScraper;
-import com.example.lazier.service.user.MemberService;
+import com.example.lazier.service.user.MyPageService;
 import java.util.Optional;
 import javax.servlet.http.HttpServletRequest;
 import lombok.AllArgsConstructor;
@@ -18,17 +18,17 @@ public class WeatherService {
 
     private final NaverWeatherScraper naverWeatherScraper;
     private final WeatherRepository weatherRepository;
-    private final MemberService memberService;
+    private final MyPageService myPageService;
 
     public void add(UserWeather userWeather) {
-        LazierUser lazierUser = memberService.searchMember(userWeather.getLazierUser().getUserId());
+        LazierUser lazierUser = myPageService.searchMember(userWeather.getLazierUser().getUserId());
         WeatherDto weatherDto = naverWeatherScraper.scrap(userWeather);
         weatherRepository.save(new Weather(lazierUser, weatherDto));
     }
 
     public WeatherDto getWeather(HttpServletRequest request) {
         long userId = Long.parseLong(request.getAttribute("userId").toString());
-        LazierUser lazierUser = memberService.searchMember(userId);
+        LazierUser lazierUser = myPageService.searchMember(userId);
 
         Optional<Weather> optionalWeather = weatherRepository.findTopByLazierUserOrderByUpdatedAtDesc(
             lazierUser);
