+++ src/main/java/com/example/lazier/service/user/MemberService.java
@@ -1,110 +1,88 @@
 package com.example.lazier.service.user;
 
 import com.example.lazier.component.MailComponents;
-import com.example.lazier.dto.user.FindPasswordRequestDto;
-import com.example.lazier.dto.user.MemberModuleResponseDto;
-import com.example.lazier.dto.user.MemberModuleUpdateRequestDto;
-import com.example.lazier.dto.user.UpdatePasswordRequestDto;
-import com.example.lazier.dto.user.MemberInfoDto;
-import com.example.lazier.exception.user.FailedFindPasswordException;
-import com.example.lazier.exception.user.NotFoundMemberException;
-import com.example.lazier.exception.user.NotMatchMemberException;
+import com.example.lazier.dto.user.MemberModuleSaveRequestDto;
+import com.example.lazier.dto.user.SignUpRequestDto;
+import com.example.lazier.dto.user.SignUpResponseDto;
+import com.example.lazier.dto.user.UniqueCodeDto;
 import com.example.lazier.persist.entity.user.LazierUser;
-import com.example.lazier.persist.repository.MemberRepository;
+import com.example.lazier.persist.entity.user.ModuleYn;
+import com.example.lazier.persist.repository.ModuleYnRepository;
 import com.example.lazier.type.MemberStatus;
-import java.util.UUID;
-import javax.servlet.http.HttpServletRequest;
+import com.example.lazier.exception.user.FailedSignUpException;
+import com.example.lazier.persist.repository.MemberRepository;
 import javax.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
-@RequiredArgsConstructor
+import java.time.LocalDateTime;
+import java.util.UUID;
+
 @Service
-@Slf4j
+@RequiredArgsConstructor
 public class MemberService {
 
-	private final MemberRepository memberRepository;
-	private final PasswordEncoder passwordEncoder;
-	private final RedisService redisService;
-	private final MailComponents mailComponents;
+    private final MemberRepository memberRepository;
+    private final ModuleYnRepository moduleYnRepository;
+    private final MailComponents mailComponents;
+    private final PasswordEncoder passwordEncoder;
 
-	public MemberModuleResponseDto main(HttpServletRequest request) {
-		LazierUser lazierUser = searchMember(parseUserId(request));
-		return MemberModuleResponseDto
-			.builder()
-			.userModuleList(lazierUser.getUserModuleList())
-			.build();
-	}
+    public SignUpResponseDto signUp(SignUpRequestDto signUpRequestDto) {
 
-	public MemberInfoDto showUserInfo(HttpServletRequest request) {
-		LazierUser lazierUser = searchMember(parseUserId(request));
-		return MemberInfoDto.of(lazierUser);
-	}
+        boolean existsEmail = memberRepository.existsByUserEmail(signUpRequestDto.getUserEmail());
+        if (existsEmail) { throw new FailedSignUpException("이미 가입된 이메일입니다."); }
 
-	@Transactional
-	public void updateUserInfo(HttpServletRequest request, MemberInfoDto memberInfoDto) {
-		LazierUser lazierUser = searchMember(parseUserId(request));
-		lazierUser.updateUserInfo(memberInfoDto);
-	}
+        LazierUser lazierUser = memberRepository.save(LazierUser.builder()
+            .userEmail(signUpRequestDto.getUserEmail())
+            .password(passwordEncoder.encode(signUpRequestDto.getPassword()))
+            .name(signUpRequestDto.getUserName())
+            .phoneNumber(signUpRequestDto.getPhoneNumber())
+            .socialType("no-social")
+            .createdAt(LocalDateTime.now())
+            .userStatus(MemberStatus.STATUS_READY.getUserStatus())
+            .emailAuthYn(false)
+            .build());
 
-	@Transactional
-	public void updatePassword(HttpServletRequest request, UpdatePasswordRequestDto passwordDto) {
-		LazierUser lazierUser = searchMember(parseUserId(request));
+        return SignUpResponseDto.builder()
+            .userId(lazierUser.getUserId().toString())
+            .build();
+    }
 
-		if (!passwordEncoder.matches(passwordDto.getPassword(), lazierUser.getPassword())) {
-			throw new NotMatchMemberException("비밀번호를 다시 입력하세요.");
-		}
-		lazierUser.setPassword(passwordEncoder.encode(passwordDto.getNewPassword()));
-	}
+    @Transactional
+    public UniqueCodeDto saveModule(MemberModuleSaveRequestDto memberModuleSaveRequestDto) {
 
-	//사용자 정보(이메일, 전화번호)가 일치하면 -> 이메일로 임시 비밀번호 전송
-	@Transactional
-	public void findPassword(HttpServletRequest request, FindPasswordRequestDto passwordDto) {
-		LazierUser lazierUser = searchMember(parseUserId(request));
+        LazierUser lazierUser = memberRepository.findByUserId(Long.parseLong(
+                memberModuleSaveRequestDto.getUserId()))
+            .orElseThrow(() -> new FailedSignUpException("회원가입을 진행하세요."));
 
-		if (!lazierUser.getUserEmail().equals(passwordDto.getUserEmail())
-			|| !lazierUser.getPhoneNumber().equals(passwordDto.getPhoneNumber())) {
-			throw new NotFoundMemberException("사용자 정보가 없습니다");
-		}
+        moduleYnRepository.save(ModuleYn.of(lazierUser, memberModuleSaveRequestDto)); //모듈 저장
 
-		String uuid = UUID.randomUUID().toString().replaceAll("-", "").substring(0, 6);
-		lazierUser.setPassword(passwordEncoder.encode(uuid));
+        String uuid = UUID.randomUUID().toString();
+        lazierUser.setEmailAuthKey(uuid);
 
-		String email = lazierUser.getUserEmail();
-		String title = "Lazier 새 비밀번호 발급";
-		String contents = "임시 비밀번호 : " + uuid;
+        String email = lazierUser.getUserEmail();
+        String title = "Lazier 가입을 축하드립니다.";
+        String contents = "아래 링크를 클릭하여 가입을 완료하세요." +
+            "<p>" +
+            "<a target='_blank' href='http://3.35.129.231:8080/user/email-auth?uuid=" + uuid + "'>가입완료</a>" +
+            "</p>";
 
-		boolean sendEmail = mailComponents.sendEmail(email, title, contents);
-		if (!sendEmail) {
-			throw new FailedFindPasswordException("메일 전송에 실패하였습니다.");
-		}
-	}
+        boolean sendEmail = mailComponents.sendEmail(email, title, contents);
+        if (!sendEmail) { throw new FailedSignUpException("메일 전송에 실패하였습니다."); }
 
-	@Transactional
-	public void withdrawal(HttpServletRequest request) {
-		LazierUser lazierUser = searchMember(parseUserId(request));
-		redisService.delValues(request);
-		//lazierUser.delete();
-		//memberRepository.delete(lazierUser);
-		lazierUser.setUserStatus(MemberStatus.STATUS_WITHDRAW.getUserStatus());
-	}
+        return UniqueCodeDto.builder() //for test
+            .uuid(uuid)
+            .build();
+    }
 
-	public LazierUser searchMember(Long userId) {
-		return memberRepository.findByUserId(userId)
-			.orElseThrow(() -> new NotFoundMemberException("사용자 정보가 없습니다."));
-	}
+    public void emailAuth(String uuid) {
+        LazierUser lazierUser = memberRepository.findByEmailAuthKey(uuid)
+                .orElseThrow(() -> new FailedSignUpException("회원가입을 진행하세요."));
 
-	public Long parseUserId(HttpServletRequest request) {
-		return Long.valueOf(request.getAttribute("userId").toString());
-	}
-
-	//모듈 업데이트
-	@Transactional
-	public void updateModule(HttpServletRequest request, MemberModuleUpdateRequestDto memberModuleUpdateRequestDto) {
-		LazierUser lazierUser = searchMember(parseUserId(request));
-		lazierUser.setUserModuleList(memberModuleUpdateRequestDto.getUserModuleList());
-	}
-
+        lazierUser.setEmailAuthYn(true);
+        lazierUser.setEmailAuthDt(LocalDateTime.now());
+        lazierUser.setUserStatus(MemberStatus.STATUS_ACTIVE.toString());
+        memberRepository.save(lazierUser);
+    }
 }
